01-ai-agents/06-crash-course/openai-docs-practice/01-agents/pyproject.toml
[project]
name = "01-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.0",
    "google-genai>=1.27.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "rich>=14.0.0",
]
